feat(2024/day18): complete part 2
Can see a fair amount of space for optimisation here. If the new byte doesn't drop into the previous
best path, we don't need to re-run the pathfinding algorithm. If it does, we would check whether
there is an alternate path that is still valid and can reach the exit.